[epub] Fix metadata for no-content elements
This fixes a bug with replaceMetadata where it replaces the wrong
element  certain metadata elements are used, and improves the functionality
of getMetadata.

The epub3 standard lists two elements of the content.opf <metadata>
element which have no content: the link element, and the OPF2 legacy
meta element.  Currently addMetadata can add these, but getMatadata
ignores them.  Since replaceMetadata uses getMetadata to find the index
of the element to replace, but then actually replaces it in the full
metadata element, it can use the wrong index.

This patch should fix the issue for all *conforming* ebooks.  If there
is a text node in the content.opf <metadata> section, then it will still
be ignored and the bug will still happen.  Fixing that probably requires
changes to the MetadataEntry type.