Change devcontainer to work with nix flake
This way it'll stay more in sync with the rest of the development
environment.  There is still some duplication of recommended vscode
extensions, but that's relatively minor.

This is a built-in devenv feature, but by default does not support nix
flakes.  We use a custom updateContentCommand to set up nix flakes
trust, handle first run devenv building, and install yarn packages.