Add devenv repo:init task before shell and tests
This should ensure the environment is set up correctly, and only run
once.  Calls after the first run will still need to be done manually, if
the task succeeds.  If the task fails (eg. due to lack of Internet
connection) the .devenv-initialized file will not be created so this
will automatically be re-run.