Fix epub metadata id attribute handling
Previously the `id` attribute was being counted twice by getMetadata, as
both the id and an attribute.  It was also being set incorrectly by
addMetadata and replaceMetadata, which added an extra `@_` to it in
addition to the one added by createXmlElement, leading to the attribute
in the XML document being `@_id` and the attribute in the parsed XML
being `@_@_id`.