Improve audio file handling
For file input:
Allow processing any type of audio file the mime type library recognizes
(as well as m4b, which it does not yet recognize).  Additionally support
audio cover.png in zip files.

For file output:
Make the default format for the opus codec be ogg instead of mp4,
because that is what the EPUB spec says is required to be supported
and I don't know of any ebook readers which support mp4/opus but not
ogg/opus.