Workaround for iOS bluetooth bug.
When using Shokz headphones, the app receives RemotePlay event instead
of RemotePause if the playback rate is faster than 1x.

This also happens with the YouTube app, so it's likely a bug in Apple's
bluetooth stack. On Android, this behavior does not happen.

Work around this by modifying the RemotePlay event handler:
Check if the player is already playing. If the player is already playing,
pause the player instead.