Fix playback speed overflowing onto multiple lines.
Replace `width` with `minWidth` in the toolbarButton style to account
for the fact that text buttons could grow beyond 40px wide.

Also, style text buttons with a bit of horizontal padding so they look
more congruent with the other buttons.

Finally, add a bit of gap between the buttons for nicer button spacing.

<!-- ps-id: 1a31fe14-157b-4f81-8e59-dcbcde144ac1 -->