Android: render StatusBar with background color from theme.
In dark mode, an unsightly white status bar mars the beauty of the app
and distracts the user. Render a StatusBar component on android with the
same background color as the rest of the app.

In the future, the app will be required by the play store to use
edge-to-edge rendering which will make the StatusBar unnecessary
but will require changes to any custom Header components that exist
in the app.

<!-- ps-id: 49f76fd5-faf3-4eb1-8073-a3b384872d79 -->